requirements.txt
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
filelock==3.9.0
Flask==2.2.0
huggingface-hub==0.12.0
idna==3.4
imap-tools==0.5.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
packaging==23.0
pandas==1.4.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
six==1.16.0
tokenizers==0.13.2
torch==1.13.0
tqdm==4.64.1
transformers==4.26.0
typing==3.7.4.3
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wincertstore==0.2
zipp==3.12.1
